Create pwmstep.py
Implements a PWM automatic heating module that allows a step down in heater efficiency to be able to control roughness of a boil. Has an initial heater setting of efficiency and time period, a setting for how many degrees before your target temp to switch to step 2, and a heater setting and time period for the second step.